remove attacks_tf (#955)
* move vatm

* remove non symbolic jsma

* mv jacobian graph and augmentation

* mv jsma symb

* mv cwl2

* mv ead

* mv deep fool

* rm lbfgs

* mv spsa

* clean up attacks tf

* autopep8

* lots of small fixes

* run tests locally

* small fixes
