+++ cleverhans/attacks/spsa.py
@@ -3,13 +3,16 @@
 import warnings
 
 import numpy as np
+from six.moves import xrange
 import tensorflow as tf
 
 from cleverhans.attacks.attack import Attack
-from cleverhans.attacks_tf import SPSAAdam, margin_logit_loss, TensorAdam
-from cleverhans.compat import reduce_mean
+from cleverhans.compat import reduce_mean, reduce_sum, reduce_max
 from cleverhans.model import Model
 from cleverhans import utils_tf
+
+tf_dtype = tf.as_dtype('float32')
+
 
 class SPSA(Attack):
   """
@@ -198,6 +201,7 @@ class SPSA(Attack):
       x_adv.append(adv_img)
     return np.concatenate(x_adv, axis=0)
 
+
 def _project_perturbation(perturbation, epsilon, input_image, clip_min=None,
                           clip_max=None):
   """Project `perturbation` onto L-infinity ball of radius `epsilon`.
@@ -221,6 +225,353 @@ def _project_perturbation(perturbation, epsilon, input_image, clip_min=None,
     new_image = utils_tf.clip_by_value(
         input_image + clipped_perturbation, clip_min, clip_max)
     return new_image - input_image
+
+
+class TensorOptimizer(object):
+  """Optimizer for Tensors rather than tf.Variables.
+
+  TensorOptimizers implement optimizers where the values being optimized
+  are ordinary Tensors, rather than Variables. TF Variables can have strange
+  behaviors when being assigned multiple times within a single sess.run()
+  call, particularly in Distributed TF, so this avoids thinking about those
+  issues. These are helper classes for the `projected_optimization`
+  method. Apart from not using Variables, they follow an interface very
+  similar to tf.Optimizer.
+  """
+
+  def _compute_gradients(self, loss_fn, x, unused_optim_state):
+    """Compute a new value of `x` to minimize `loss_fn`.
+
+    Args:
+        loss_fn: a callable that takes `x`, a batch of images, and returns
+            a batch of loss values. `x` will be optimized to minimize
+            `loss_fn(x)`.
+        x: A list of Tensors, the values to be updated. This is analogous
+            to the `var_list` argument in standard TF Optimizer.
+        unused_optim_state: A (possibly nested) dict, containing any state
+            info needed for the optimizer.
+
+    Returns:
+        new_x: A list of Tensors, the same length as `x`, which are updated
+        new_optim_state: A dict, with the same structure as `optim_state`,
+            which have been updated.
+    """
+
+    # Assumes `x` is a list,
+    # and contains a tensor representing a batch of images
+    assert len(x) == 1 and isinstance(x, list), \
+        'x should be a list and contain only one image tensor'
+    x = x[0]
+    loss = reduce_mean(loss_fn(x), axis=0)
+    return tf.gradients(loss, x)
+
+  def _apply_gradients(self, grads, x, optim_state):
+    raise NotImplementedError(
+        "_apply_gradients should be defined in each subclass")
+
+  def minimize(self, loss_fn, x, optim_state):
+    grads = self._compute_gradients(loss_fn, x, optim_state)
+    return self._apply_gradients(grads, x, optim_state)
+
+  def init_optim_state(self, x):
+    """Returns the initial state of the optimizer.
+
+    Args:
+        x: A list of Tensors, which will be optimized.
+
+    Returns:
+        A dictionary, representing the initial state of the optimizer.
+    """
+    raise NotImplementedError(
+        "init_optim_state should be defined in each subclass")
+
+
+class TensorGradientDescent(TensorOptimizer):
+  """Vanilla Gradient Descent TensorOptimizer."""
+
+  def __init__(self, lr):
+    self._lr = lr
+
+  def init_state(self, x):
+    return {}
+
+  def _apply_gradients(self, grads, x, optim_state):
+    new_x = [None] * len(x)
+    for i in xrange(len(x)):
+      new_x[i] = x[i] - self._lr * grads[i]
+    return new_x, optim_state
+
+
+class TensorAdam(TensorOptimizer):
+  """The Adam optimizer defined in https://arxiv.org/abs/1412.6980."""
+
+  def __init__(self, lr=0.001, beta1=0.9, beta2=0.999, epsilon=1e-9):
+    self._lr = lr
+    self._beta1 = beta1
+    self._beta2 = beta2
+    self._epsilon = epsilon
+
+  def init_state(self, x):
+    optim_state = {}
+    optim_state["t"] = 0.
+    optim_state["m"] = [tf.zeros_like(v) for v in x]
+    optim_state["u"] = [tf.zeros_like(v) for v in x]
+    return optim_state
+
+  def _apply_gradients(self, grads, x, optim_state):
+    """Refer to parent class documentation."""
+    new_x = [None] * len(x)
+    new_optim_state = {
+        "t": optim_state["t"] + 1.,
+        "m": [None] * len(x),
+        "u": [None] * len(x)
+    }
+    t = new_optim_state["t"]
+    for i in xrange(len(x)):
+      g = grads[i]
+      m_old = optim_state["m"][i]
+      u_old = optim_state["u"][i]
+      new_optim_state["m"][i] = (
+          self._beta1 * m_old + (1. - self._beta1) * g)
+      new_optim_state["u"][i] = (
+          self._beta2 * u_old + (1. - self._beta2) * g * g)
+      m_hat = new_optim_state["m"][i] / (1. - tf.pow(self._beta1, t))
+      u_hat = new_optim_state["u"][i] / (1. - tf.pow(self._beta2, t))
+      new_x[i] = (
+          x[i] - self._lr * m_hat / (tf.sqrt(u_hat) + self._epsilon))
+    return new_x, new_optim_state
+
+
+class SPSAAdam(TensorAdam):
+  """Optimizer for gradient-free attacks in https://arxiv.org/abs/1802.05666.
+
+  Gradients estimates are computed using Simultaneous Perturbation Stochastic
+  Approximation (SPSA), combined with the ADAM update rule.
+  """
+
+  def __init__(self,
+               lr=0.01,
+               delta=0.01,
+               num_samples=128,
+               num_iters=1,
+               compare_to_analytic_grad=False):
+    super(SPSAAdam, self).__init__(lr=lr)
+    assert num_samples % 2 == 0, "number of samples must be even"
+    self._delta = delta
+    self._num_samples = num_samples // 2  # Since we mirror +/- delta later
+    self._num_iters = num_iters
+    self._compare_to_analytic_grad = compare_to_analytic_grad
+
+  def _get_delta(self, x, delta):
+    x_shape = x.get_shape().as_list()
+    delta_x = delta * tf.sign(
+        tf.random_uniform(
+            [self._num_samples] + x_shape[1:],
+            minval=-1.,
+            maxval=1.,
+            dtype=tf_dtype))
+    return delta_x
+
+  def _compute_gradients(self, loss_fn, x, unused_optim_state):
+    """Compute gradient estimates using SPSA."""
+    # Assumes `x` is a list, containing a [1, H, W, C] image
+    # If static batch dimension is None, tf.reshape to batch size 1
+    # so that static shape can be inferred
+    assert len(x) == 1
+    static_x_shape = x[0].get_shape().as_list()
+    if static_x_shape[0] is None:
+      x[0] = tf.reshape(x[0], [1] + static_x_shape[1:])
+    assert x[0].get_shape().as_list()[0] == 1
+    x = x[0]
+    x_shape = x.get_shape().as_list()
+
+    def body(i, grad_array):
+      delta = self._delta
+      delta_x = self._get_delta(x, delta)
+      delta_x = tf.concat([delta_x, -delta_x], axis=0)
+      loss_vals = tf.reshape(
+          loss_fn(x + delta_x),
+          [2 * self._num_samples] + [1] * (len(x_shape) - 1))
+      avg_grad = reduce_mean(loss_vals * delta_x, axis=0) / delta
+      avg_grad = tf.expand_dims(avg_grad, axis=0)
+      new_grad_array = grad_array.write(i, avg_grad)
+      return i + 1, new_grad_array
+
+    def cond(i, _):
+      return i < self._num_iters
+
+    _, all_grads = tf.while_loop(
+        cond,
+        body,
+        loop_vars=[
+            0, tf.TensorArray(size=self._num_iters, dtype=tf_dtype)
+        ],
+        back_prop=False,
+        parallel_iterations=1)
+    avg_grad = reduce_sum(all_grads.stack(), axis=0)
+    return [avg_grad]
+
+
+def margin_logit_loss(model_logits, label, nb_classes=10, num_classes=None):
+  """Computes difference between logit for `label` and next highest logit.
+
+  The loss is high when `label` is unlikely (targeted by default).
+  This follows the same interface as `loss_fn` for TensorOptimizer and
+  projected_optimization, i.e. it returns a batch of loss values.
+  """
+  if num_classes is not None:
+    warnings.warn("`num_classes` is depreciated. Switch to `nb_classes`."
+                  " `num_classes` may be removed on or after 2019-04-23.")
+    nb_classes = num_classes
+    del num_classes
+  if 'int' in str(label.dtype):
+    logit_mask = tf.one_hot(label, depth=nb_classes, axis=-1)
+  else:
+    logit_mask = label
+  if 'int' in str(logit_mask.dtype):
+    logit_mask = tf.to_float(logit_mask)
+  try:
+    label_logits = reduce_sum(logit_mask * model_logits, axis=-1)
+  except TypeError:
+    raise TypeError("Could not take row-wise dot product between "
+                    "logit mask, of dtype " + str(logit_mask.dtype)
+                    + " and model_logits, of dtype "
+                    + str(model_logits.dtype))
+  logits_with_target_label_neg_inf = model_logits - logit_mask * 99999
+  highest_nonlabel_logits = reduce_max(
+      logits_with_target_label_neg_inf, axis=-1)
+  loss = highest_nonlabel_logits - label_logits
+  return loss
+
+
+def _apply_black_border(x, border_size):
+  orig_height = x.get_shape().as_list()[1]
+  orig_width = x.get_shape().as_list()[2]
+  x = tf.image.resize_images(x, (orig_width - 2*border_size,
+                                 orig_height - 2*border_size))
+
+  return tf.pad(x, [[0, 0],
+                    [border_size, border_size],
+                    [border_size, border_size],
+                    [0, 0]], 'CONSTANT')
+
+
+def _apply_transformation(inputs):
+  x, trans = inputs[0], inputs[1]
+  dx, dy, angle = trans[0], trans[1], trans[2]
+  height = x.get_shape().as_list()[1]
+  width = x.get_shape().as_list()[2]
+
+  # Pad the image to prevent two-step rotation / translation from truncating
+  # corners
+  max_dist_from_center = np.sqrt(height**2+width**2) / 2
+  min_edge_from_center = float(np.min([height, width])) / 2
+  padding = np.ceil(max_dist_from_center -
+                    min_edge_from_center).astype(np.int32)
+  x = tf.pad(x, [[0, 0],
+                 [padding, padding],
+                 [padding, padding],
+                 [0, 0]],
+             'CONSTANT')
+
+  # Apply rotation
+  angle *= np.pi / 180
+  x = tf.contrib.image.rotate(x, angle, interpolation='BILINEAR')
+
+  # Apply translation
+  dx_in_px = -dx * height
+  dy_in_px = -dy * width
+  translation = tf.convert_to_tensor([dx_in_px, dy_in_px])
+
+  try:
+    x = tf.contrib.image.translate(x, translation, interpolation='BILINEAR')
+  except AttributeError as e:
+    print("WARNING: SpatialAttack requires tf 1.6 or higher")
+    raise e
+  x = tf.contrib.image.translate(x, translation, interpolation='BILINEAR')
+  return tf.image.resize_image_with_crop_or_pad(x, height, width)
+
+
+def spm(x, model, y=None, n_samples=None, dx_min=-0.1,
+        dx_max=0.1, n_dxs=5, dy_min=-0.1, dy_max=0.1, n_dys=5,
+        angle_min=-30, angle_max=30, n_angles=31, black_border_size=0):
+  """
+  TensorFlow implementation of the Spatial Transformation Method.
+  :return: a tensor for the adversarial example
+  """
+  if y is None:
+    preds = model.get_probs(x)
+    # Using model predictions as ground truth to avoid label leaking
+    preds_max = reduce_max(preds, 1, keepdims=True)
+    y = tf.to_float(tf.equal(preds, preds_max))
+    y = tf.stop_gradient(y)
+    del preds
+  y = y / reduce_sum(y, 1, keepdims=True)
+
+  # Define the range of transformations
+  dxs = np.linspace(dx_min, dx_max, n_dxs)
+  dys = np.linspace(dy_min, dy_max, n_dys)
+  angles = np.linspace(angle_min, angle_max, n_angles)
+
+  if n_samples is None:
+    import itertools
+    transforms = list(itertools.product(*[dxs, dys, angles]))
+  else:
+    sampled_dxs = np.random.choice(dxs, n_samples)
+    sampled_dys = np.random.choice(dys, n_samples)
+    sampled_angles = np.random.choice(angles, n_samples)
+    transforms = zip(sampled_dxs, sampled_dys, sampled_angles)
+  transformed_ims = parallel_apply_transformations(
+      x, transforms, black_border_size)
+
+  def _compute_xent(x):
+    preds = model.get_logits(x)
+    return tf.nn.softmax_cross_entropy_with_logits_v2(
+        labels=y, logits=preds)
+
+  all_xents = tf.map_fn(
+      _compute_xent,
+      transformed_ims,
+      parallel_iterations=1)  # Must be 1 to avoid keras race conditions
+
+  # Return the adv_x with worst accuracy
+
+  # all_xents is n_total_samples x batch_size (SB)
+  all_xents = tf.stack(all_xents)  # SB
+
+  # We want the worst case sample, with the largest xent_loss
+  worst_sample_idx = tf.argmax(all_xents, axis=0)  # B
+
+  batch_size = tf.shape(x)[0]
+  keys = tf.stack([
+      tf.range(batch_size, dtype=tf.int32),
+      tf.cast(worst_sample_idx, tf.int32)
+  ], axis=1)
+  transformed_ims_bshwc = tf.einsum('sbhwc->bshwc', transformed_ims)
+  after_lookup = tf.gather_nd(transformed_ims_bshwc, keys)  # BHWC
+  return after_lookup
+
+
+def parallel_apply_transformations(x, transforms, black_border_size=0):
+  transforms = tf.convert_to_tensor(transforms, dtype=tf.float32)
+  x = _apply_black_border(x, black_border_size)
+
+  num_transforms = transforms.get_shape().as_list()[0]
+  im_shape = x.get_shape().as_list()[1:]
+
+  # Pass a copy of x and a transformation to each iteration of the map_fn
+  # callable
+  tiled_x = tf.reshape(
+      tf.tile(x, [num_transforms, 1, 1, 1]),
+      [num_transforms, -1] + im_shape)
+  elems = [tiled_x, transforms]
+  transformed_ims = tf.map_fn(
+      _apply_transformation,
+      elems,
+      dtype=tf.float32,
+      parallel_iterations=1,  # Must be 1 to avoid keras race conditions
+  )
+  return transformed_ims
 
 
 def projected_optimization(loss_fn,
@@ -322,7 +673,8 @@ def projected_optimization(loss_fn,
           loss = tf.Print(loss, [loss], "Total batch loss")
 
       if early_stop:
-        i = tf.cond(tf.less(loss, early_stop_loss_threshold), lambda: float(num_steps), lambda: i)
+        i = tf.cond(tf.less(loss, early_stop_loss_threshold),
+                    lambda: float(num_steps), lambda: i)
 
     return i + 1, projected_perturbation, nest.flatten(new_optim_state)
