+++ cleverhans/attacks/virtual_adversarial_method.py
@@ -4,9 +4,14 @@
 
 import warnings
 
+import tensorflow as tf
+
 from cleverhans.attacks.attack import Attack
 from cleverhans.model import Model, CallableModelWrapper
 from cleverhans.model import wrapper_warning_logits
+from cleverhans import utils_tf
+
+tf_dtype = tf.as_dtype('float32')
 
 class VirtualAdversarialMethod(Attack):
   """
@@ -98,24 +103,23 @@ class VirtualAdversarialMethod(Attack):
                     "2019-04-26.")
     return True
 
+
 def vatm(model,
          x,
          logits,
          eps,
-         back='tf',
          num_iterations=1,
          xi=1e-6,
          clip_min=None,
-         clip_max=None):
+         clip_max=None,
+         scope=None):
   """
-  A wrapper for the perturbation methods used for virtual adversarial
-  training : https://arxiv.org/abs/1507.00677
-  It calls the right function, depending on the
-  user's backend.
-
+  Tensorflow implementation of the perturbation method used for virtual
+  adversarial training: https://arxiv.org/abs/1507.00677
   :param model: the model which returns the network unnormalized logits
   :param x: the input placeholder
-  :param logits: the model's unnormalized output tensor
+  :param logits: the model's unnormalized output tensor (the input to
+                 the softmax layer)
   :param eps: the epsilon (input variation parameter)
   :param num_iterations: the number of iterations
   :param xi: the finite difference parameter
@@ -123,18 +127,19 @@ def vatm(model,
                   value for components of the example returned
   :param clip_max: optional parameter that can be used to set a maximum
                   value for components of the example returned
+  :param seed: the seed for random generator
   :return: a tensor for the adversarial example
-
   """
-  assert back == 'tf'
-  # Compute VATM using TensorFlow
-  from cleverhans.attacks_tf import vatm as vatm_tf
-  return vatm_tf(
-      model,
-      x,
-      logits,
-      eps,
-      num_iterations=num_iterations,
-      xi=xi,
-      clip_min=clip_min,
-      clip_max=clip_max)
+  with tf.name_scope(scope, "virtual_adversarial_perturbation"):
+    d = tf.random_normal(tf.shape(x), dtype=tf_dtype)
+    for _ in range(num_iterations):
+      d = xi * utils_tf.l2_batch_normalize(d)
+      logits_d = model.get_logits(x + d)
+      kl = utils_tf.kl_with_logits(logits, logits_d)
+      Hd = tf.gradients(kl, d)[0]
+      d = tf.stop_gradient(Hd)
+    d = eps * utils_tf.l2_batch_normalize(d)
+    adv_x = x + d
+    if (clip_min is not None) and (clip_max is not None):
+      adv_x = tf.clip_by_value(adv_x, clip_min, clip_max)
+    return adv_x
